Move dashboard to /
