+++ apps/dashboard/src/app.postcss
@@ -2,5 +2,4 @@
 html,
 body {
 	@apply h-full overflow-hidden;
-	@apply bg-surface-50-900-token;
 }
