+++ apps/dashboard/src/routes/+page.svelte
@@ -1,4 +1,30 @@
 <script lang="ts">
+	import { faAdd } from '@fortawesome/free-solid-svg-icons';
+	import Fa from 'svelte-fa';
+
+	import { modalStore } from '@skeletonlabs/skeleton';
+	import type { ModalSettings } from '@skeletonlabs/skeleton';
+	import CreateURLPrefixModal from './CreateURLPrefixModal.svelte';
+
+	const urlCreationModal: ModalSettings = {
+		type: 'component',
+		component: { ref: CreateURLPrefixModal },
+		response: async (url: string) => {}
+	};
 </script>
 
-<div class="sm:container sm:mx-auto justify-center p-4">Hello World</div>
+<div class="sm:container sm:mx-auto justify-center p-8">
+	<div class="flex justify-between">
+		<div class="flex text-4xl">Dashboard</div>
+		<button
+			type="button"
+			class="btn variant-filled-surface"
+			on:click={() => {
+				modalStore.trigger(urlCreationModal);
+			}}
+		>
+			<Fa icon={faAdd} />
+			<span>Create URL Prefix</span>
+		</button>
+	</div>
+</div>
