+++ apps/dashboard/src/routes/CreateURLPrefixModal.svelte
@@ -7,11 +7,9 @@
 	import { modalStore } from '@skeletonlabs/skeleton';
 	import { Stepper, Step } from '@skeletonlabs/skeleton';
 
-	// Step 1 (domain)
 	let domain = '';
 	$: isDomainValid = domain.length != 0; // TODO: Validate more.
 
-	// Step 2 (finish)
 	function onFormSubmit(event: Event): void {
 		if ($modalStore[0].response) $modalStore[0].response(domain);
 		modalStore.close();
